// src/registry/openapi/models/EndpointResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EndpointResponse = {
    /**
     * Tenant ID
     */
    tenantId: string;
    /**
     * Auth service URL
     */
    authUrl: string;
    /**
     * Server URL
     */
    serverUrl: string;
    /**
     * Web URL
     */
    webUrl: string;
    /**
     * API version
     */
    apiVersion: string;
};